// src/app/interfaces/update-companion.ts
export interface UpdateCompanion {
  idPaciente: number;
  codigoUbigeo: string;
  apellidoMaterno: string;
  fechaNacimiento: Date;
  idTipoDocumentoIdentidad: number;
  nombres: string;
  idParentesco: number;
  numeroDocumentoIdentidad: string;
  direccion: string;
  telefonoContacto: string;
  apellidoPaterno: string;
}